[refactor]: 타임캡슐 구조 리팩터링
