+++ app/(public)/timecapsule/layout.tsx
@@ -1,5 +1,5 @@
 import { PropsWithChildren } from 'react';
 
 export default function TimeCapsuleLayout({ children }: PropsWithChildren) {
-  return <div className="fixed inset-0 h-dvh w-dvw z-50">{children}</div>;
+  return <div className="fixed inset-0 h-dvh w-dvw z-50 bg-black">{children}</div>;
 }
